+++ spec/spec_helper.rb
@@ -30,6 +30,10 @@ RSpec.shared_context 'ruby 2.7', :ruby27 do
   let(:ruby_version) { 2.7 }
 end
 
+RSpec.shared_context 'ruby 2.8', :ruby28 do
+  let(:ruby_version) { 2.8 }
+end
+
 module DefaultRubyVersion
   extend RSpec::SharedContext